Fix #19, Combine preprocess with input check and save files as .npy
