+++ Paco_classifier/__init__.py
@@ -1,5 +1,4 @@
 from . import training_engine_sae
-from . import input_settings_test
 
 from . import recognition_engine
 from . import fast_trainer_lib
